refactor(mcp): extract shared MacroConfig + validators + executor builders into fuzz_macro_common.go
Pure refactor — extracted shared fuzz-macro-hook infrastructure from
internal/mcp/fuzz_http_helpers.go into internal/mcp/fuzz_macro_common.go
so USK-984 (fuzz_ws) / USK-985 (fuzz_grpc) / USK-986 (fuzz_raw) can reuse it.

Extracted symbols (all exported, package mcp):

- MacroConfig (renamed from fuzzHTTPMacroConfig — all json/jsonschema tags
  preserved verbatim)
- ValidateMacroConfig
- BuildIterationHookExecutor + resolveRunInterval (unexported helper)
- BuildJobHookExecutor
- InjectVarsRespectingReserved
- BumpHookIterationCount (converted from method on *fuzzHTTPVariantLoop to
  free function)
- IsJobScope / IsIterationScope (pure predicates over *MacroConfig — needed
  by USK-984/985/986 even though not listed in the Issue's extraction set)
- PrepareIteration (extracted kvStore-build prefix of prepareIteration —
  jobKV copy + pre/post Vars injection + connector.SeedIterationKV)

No behavior change in fuzz_http — all existing tests pass unchanged.
The remaining fuzz_http-specific variant-loop wiring (fuzzHTTPVariantLoop
struct + methods, prepareIteration's upstream-proxy and pre-macro dispatch
suffix) stays in fuzz_http_helpers.go because those touch protocol-specific
row recording.

Refs: USK-983
